Proyecto/pizzapi/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
psycopg2
asyncpg
python-dotenv
pydantic[email]
passlib
alembic
